requirements.txt
altair
attrs
backports.zoneinfo
blinker
cachetools
certifi
charset-normalizer
click
colorama
cycler
decorator
fonttools
gitdb
GitPython
idna
importlib-metadata
importlib-resources
Jinja2
jsonschema
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
mdurl
numpy
opencv-python
opencv-python-headless
packaging
pandas
Pillow
pkgutil_resolve_name
protobuf
psutil
py-cpuinfo
pyarrow
pydeck
Pygments
Pympler
pyparsing
pyrsistent
python-dateutil
pytz
pytz-deprecation-shim
pywin32
PyYAML
requests
rich
scipy
seaborn
six
smmap
streamlit
tenacity
toml
toolz
tornado
tqdm
typing_extensions
tzdata
tzlocal
ultralytics
urllib3
validators
watchdog
zipp
